% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.mlc_object}
\alias{.mlc_object}
\title{Helper function for building JSON objects}
\usage{
.mlc_object(...)
}
\arguments{
\item{...}{key-value pairs with JSON strings for values}
}
\description{
Helper function for building JSON objects
}
