% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.mlc_record}
\alias{.mlc_record}
\title{Helper function for building record types}
\usage{
.mlc_record(...)
}
\arguments{
\item{...}{key-value pairs with JSON strings for values}
}
\description{
Helper function for building record types
}
