% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.mlc_table}
\alias{.mlc_table}
\title{Helper function for building table types}
\usage{
.mlc_table(...)
}
\arguments{
\item{...}{key-value pairs with JSON strings for column types}
}
\description{
Helper function for building table types
}
