% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.mlc_tuple}
\alias{.mlc_tuple}
\title{Helper function for building tuple types}
\usage{
.mlc_tuple(...)
}
\arguments{
\item{...}{elements of the tuple as JSON strings}
}
\description{
A tuple corresponds to an unnamed R list.
}
