% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mlc_deserialize_}
\alias{mlc_deserialize_}
\title{Deserialize JSON data to an R native type using a JSON type string}
\usage{
mlc_deserialize_(data, schema)
}
\arguments{
\item{data}{JSON data}

\item{schema}{JSON type string}
}
\description{
Deserialize JSON data to an R native type using a JSON type string
}
